+++ backend/src/Responder/Admin/UsersResponder.php
@@ -2,8 +2,8 @@
 
 namespace App\Responder\Admin;
 
-use App\MainApp;
 use App\Responder\ResponderInterface;
+use App\Security\Authentication\Authentication;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -12,6 +12,6 @@ class UsersResponder implements ResponderInterface
     public function respond(ServerRequestInterface $request, ResponseInterface $response, $args): ResponseInterface
     {
         // The user authenticated with request header access token
-        dd(MainApp::getInstance()->getUser());
+        dd(Authentication::getInstance()->getUser());
     }
 }
